♻️ refactor: 메뉴 옵션 컴포넌트 분리 (#26)
