+++ src/components/menu-item/styles/menu-item-form.module.css
@@ -104,10 +104,9 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  width: 40px;
-  height: 40px;
+  width: 4rem;
+  height: 4rem;
   border-radius: 50%;
-  color: #fff;
   background-color: var(--point-color);
   font-size: 2rem;
 }
